// ssyweb/SSYNEW/src/app/claim/models/payinslip-attachment.ts
export class PayInSlipAttachmentModel {
    payInSlipId: number;
    pfDepositId: number;
    attachmentType: number;
    fileName: string;
    filePath: string;
    fileContent: any;
    statusId: number;
    createdBy: number;
    attachmenturl: any;
    stringContent: string;
    attachmentTypeString: string;

    bankDepositDate: Date;
    scrollNumber: number;
}